dom/tests/mochitest/pointerlock/file_DOMtree.html

<!DOCTYPE HTML>
<html>
  <!--
  https://bugzilla.mozilla.org/show_bug.cgi?id=633602

  Test DOM tree in full screen
  -->
  <head>
    <title>Test for Bug 633602</title>
    <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js">
    </script>
    <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js">
    </script>
    <script type="application/javascript" src="mouselock_util.js"></script>
    <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
    <style>
    </style>
  </head>
  <body onload="start();">
    <a target="_blank"
       href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">
      Mozilla Bug 633602
    </a>
    <pre id="test">
      <script type="text/javascript">
        /*
         * Test for Bug 633602 
         * Checks if element being locked is attached to the DOM Tree
         * Also checks if pointer is unlocked when element is removed from
         * the DOM Tree
         */
        SimpleTest.waitForExplicitFinish();
        var pointer = navigator.mozPointer;
        var canvas = document.createElement('canvas');

        document.addEventListener("mozfullscreenchange", function (e) {
          if (document.mozFullScreen) {
            pointer.lock(canvas,function() {
              /*
               * If element is attached to the DOM Tree, pointer should be
               * locked
               */
              ok(pointer.isLocked, "Mouse should be locked when DOM element " +
                "is in the tree.");

              /*
               * If the element is removed from the DOM Tree, pointer should
               * be unlocked
               */
              document.body.removeChild(canvas);
              is(pointer.isLocked, false, "Pointer should be unlocked when " +
                "an element is removed the DOM Tree");
              document.mozCancelFullScreen();
            }, function () {
              /*
               * Failure callback shouldn't fire when element is attached to
               * a DOM Tree and in fullscreen mode
               */
              ok(pointer.isLocked, "Mouse should be locked when DOM " +
                "element is in the tree and failure callback should NOT be " +
                "dispatched");
              document.mozCancelFullScreen();
            });
          } else {
            SimpleTest.finish();
          }
        }, false);

        function start() {
          SimpleTest.waitForFocus(function() {
            pointer.lock(canvas, function () {
                /*
                 * Success callback shouldn't fire, since DOM is not attached
                 * to a DOM Tree
                 */
                ok( false, "If an element is NOT in the DOM " +
                  "Tree pointer should NOT be locked and success callback " +
                  "should NOT be dispatched");
                SimpleTest.finish();
              }, function () {
                /*
                 * Element is not attached to a DOM Tree, pointer shouldn't be
                 * locked
                 */
                is(pointer.isLocked, false, "If an element is NOT in the " +
                  "DOM Tree pointer should NOT be locked");
            });
            document.body.appendChild(canvas);
            canvas.mozRequestFullScreen();
          });
        }
      </script>
    </pre>
  </body>
</html>
